+++ demisto_sdk/commands/doc_reviewer/doc_reviewer.py
@@ -52,6 +52,10 @@ class DocReviewer:
 
         if release_notes_only:
             self.SUPPORTED_FILE_TYPES = [FileType.RELEASE_NOTES]
+            # if running doc-review --release-notes there is no need to consider invalid schema files of yml/json
+            self.ignore_invalid_schema_file = True
+        else:
+            self.ignore_invalid_schema_file = False
 
         self.files = set()  # type:Set
         self.spellchecker = SpellChecker()
@@ -85,7 +89,9 @@ class DocReviewer:
         for root, _, files in os.walk(dir_name):
             for file_name in files:
                 full_path = (os.path.join(root, file_name))
-                if find_type(full_path) in self.SUPPORTED_FILE_TYPES:
+                if find_type(
+                    full_path, ignore_invalid_schema_file=self.ignore_invalid_schema_file
+                ) in self.SUPPORTED_FILE_TYPES:
                     self.files.add(str(full_path))
 
     def gather_all_changed_files(self):
@@ -99,7 +105,9 @@ class DocReviewer:
         click.secho('Gathering all changed files from git', fg='bright_cyan')
         for file in self.gather_all_changed_files():
             file = str(file)
-            if os.path.isfile(file) and find_type(file) in self.SUPPORTED_FILE_TYPES:
+            if os.path.isfile(file) and find_type(
+                file, ignore_invalid_schema_file=self.ignore_invalid_schema_file
+            ) in self.SUPPORTED_FILE_TYPES:
                 self.files.add(file)
 
     def get_files_to_run_on(self):
@@ -110,7 +118,9 @@ class DocReviewer:
         elif os.path.isdir(self.file_path):
             self.get_all_md_and_yml_files_in_dir(self.file_path)
 
-        elif find_type(self.file_path) in self.SUPPORTED_FILE_TYPES:
+        elif find_type(
+            self.file_path, ignore_invalid_schema_file=self.ignore_invalid_schema_file
+        ) in self.SUPPORTED_FILE_TYPES:
             self.files.add(self.file_path)
 
     def print_unknown_words(self):